+++ CHANGELOG.md
@@ -4,6 +4,8 @@ FlatLaf Change Log
 ## 3.6.1-SNAPSHOT
 
 - Extras: Support JSVG 2.0.0. Minimum JSVG version is now 1.6.0. (issue #997)
+- ToggleButton: Styling `selectedForeground` did not work if `foreground` is
+  also styled. (issue #1017)
 - JideSplitButton: Fixed updating popup when switching theme. (issue #1000)
 - IntelliJ Themes: Fixed logging false errors when loading 3rd party
   `.theme.json` files. (issue #990)